Move DownloadError into shared project
